requirements.txt
absl-py==0.11.0
astunparse==1.6.3
atari-py==0.2.6
attrs==20.3.0
-e git+git@github.com:howardh/rl.git@f898315844ccb0af6e03f664d66d318cf626c503#egg=bsuite
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
chex==0.0.3
click==7.1.2
cloudpickle==1.6.0
cycler==0.10.0
Cython==0.29.23
decorator==4.4.2
descartes==1.1.0
dill==0.3.3
dm-env==1.3
dm-haiku==0.0.3
dm-tree==0.1.5
-e git+git@github.com:howardh/experiment.git@cb619a933da9e02036f1ab615034c188150f65af#egg=experiment
fabulous==0.3.0
fasteners==0.16
flatbuffers==1.12
frozendict==1.2
future==0.18.2
gast==0.3.3
gitdb==4.0.7
GitPython==3.1.18
glfw==2.1.0
google-auth==1.28.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
gym==0.17.3
# Editable install with no version control (gym-classification==0.0.1)
-e /home/howard/Code/gym-classification
-e git+ssh://git@github.com/howardh/gym-fourrooms.git@4576383f830a7d80e75a83dc8efc25b0ffee4280#egg=gym_fourrooms
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.4.0
iniconfig==1.1.1
jax==0.2.9
jaxlib==0.1.60
joblib==1.0.1
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
levenshtein==0.12.0
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mizani==0.7.2
mujoco-py==2.0.2.13
networkx==2.5
numpy==1.20.2
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
optax==0.0.2
packaging==20.9
palettable==3.3.0
pandas==1.2.1
patsy==0.5.1
pdoc3==0.9.2
Pillow==7.2.0
plotnine==0.7.1
pluggy==0.13.1
protobuf==3.15.8
py==1.10.0
pyaml==20.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==3.1.7
pycparser==2.20
pyglet==1.5.0
pyparsing==2.4.7
pytest==6.2.2
pytest-mock==3.6.1
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
-e git+git@github.com:howardh/rl.git@f898315844ccb0af6e03f664d66d318cf626c503#egg=rl
rlax==0.0.3
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.1
scikit-optimize==0.8.1
scipy==1.6.0
simple-slurm==0.1.7
six==1.15.0
sklearn==0.0
smmap==4.0.0
stable-baselines3==1.0
statsmodels==0.12.2
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2021.2.1
toml==0.10.2
toolz==0.11.1
torch==1.9.0
torchvision==0.10.0
tqdm==4.56.0
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
